client/examples/llm/requirements.txt
# Additional dependencies for LLM examples

# For OpenAI (llm.py)
openai>=1.0.0
python-dotenv>=1.0.0

# For Ollama (llm_ollama.py)
requests>=2.31.0
python-dotenv>=1.0.0

numpy>=1.24.0
